+++ 01-MCAL/04-NVIC_DRIVER/nvic.h
@@ -173,27 +173,27 @@ void NVIC_voidInitPriority (void);
  * Input:  Interrupt request number NVIC_IRQx , PriorityLevel
  * Output: void
  * */
-void NVIC_u8SetPriority(u8 Copy_u8IntReqNum , u8 Copy_u8PriorityLevel);
+void NVIC_u8SetPriority(u32 Copy_u8IntReqNum , u8 Copy_u8PriorityLevel);
 
 
 /*Description : This API shall Enable interrupt
  * Input: Interrupt request number NVIC_IRQx
  * output: Error_s  */
-ERROR_STATUS NVIC_u8EnableIRQ (u8 Copy_u8IntReqNum);
+ERROR_STATUS NVIC_u8EnableIRQ (u32 Copy_u8IntReqNum);
 
 
 
 /*Description : This API shall Disable interrupt
  * Input: Interrupt request number NVIC_IRQx
  * output: Error_s  */
-ERROR_STATUS NVIC_u8DisbaleIRQ (u8 Copy_u8IntReqNum);
+ERROR_STATUS NVIC_u8DisbaleIRQ (u32 Copy_u8IntReqNum);
 
 
 
 /*Description : This API shall get active status of interrupt
  * Input: Interrupt request number NVIC_IRQx
  * output: u8 active status zero if not active and larger than 1 if active */
-u8 NVIC_u8IsActive (u8 Copy_u8IntReqNum);
+u8 NVIC_u8IsActive (u32 Copy_u8IntReqNum);
 
 
 
@@ -201,7 +201,7 @@ u8 NVIC_u8IsActive (u8 Copy_u8IntReqNum);
  * Input: Interrupt request number NVIC_IRQx
  * output: u8 priority value */
 
-u8 NVIC_u8GetPriority (u8 Copy_u8IntReqNum);
+u8 NVIC_u8GetPriority (u32 Copy_u8IntReqNum);
 
 
 
@@ -209,7 +209,7 @@ u8 NVIC_u8GetPriority (u8 Copy_u8IntReqNum);
  * Input: Interrupt request number NVIC_IRQx
  * output: u8 ERROR STATUS*/
 
-ERROR_STATUS NVIC_u8SWIntGenerate(u8 Copy_u8IntReqNum);  //NVIC_SetPendingIRQ(3);NVIC->STIR = 3;
+ERROR_STATUS NVIC_u8SWIntGenerate(u32 Copy_u8IntReqNum);  //NVIC_SetPendingIRQ(3);NVIC->STIR = 3;
 
 
 
@@ -217,7 +217,7 @@ ERROR_STATUS NVIC_u8SWIntGenerate(u8 Copy_u8IntReqNum);  //NVIC_SetPendingIRQ(3)
  * Input: Interrupt request number NVIC_IRQx
  * output: u8 ERROR STATUS*/
 
-ERROR_STATUS NVIC_u8ClearPending (u8 Copy_u8IntReqNum);
+ERROR_STATUS NVIC_u8ClearPending (u32 Copy_u8IntReqNum);
